# src/openapi.yaml
openapi: 3.1.0
info:
  title: Time-addressable Media Store
  description: 'The Time-addressable Media Store (TAMS) is used for storing segmented
    media Flows.

    The TAMS consists of one or more Storage Backends for the media Flow Segment Objects
    and a service providing a database index of the Flow Segments.

    This document is a specification of the service API.


    See the [bbc/tams](https://github.com/bbc/tams) repository for more background
    on TAMS.


    **Note**: the examples provided in this specification are for a system which uses
    a Storage Backend that provides HTTP URLs for uploading and downloading Media
    Objects in buckets.

    This could for example be implemented using an AWS S3 compatible store using presigned
    URLs or by a simple file system storage with an HTTP frontend.

    Clients should parse the [/service](#/operations/GET_service) endpoint and handle
    the Storage Backend type in use appropriately.

    '
  version: '8.0'
  contact:
    name: BBC R&D - Cloud-Fit Production Team
    email: cloudfit-opensource@rd.bbc.co.uk
    url: https://github.com/bbc/tams
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
- url: http://localhost:4010
  description: Local mock of API
- url: https://example.com/tams/{version}
  description: Example TAMS service
  variables:
    version:
      description: API version
      default: v8.0
security:
- bearer_token_auth: []
- basic_auth: []
paths:
...
  /service/webhooks:
    head:
      summary: List Webhook URLs
      description: Return webhooks path headers
      operationId: HEAD_webhooks
      tags:
      - Webhooks
      parameters:
      - name: tag.{name}
        in: query
        description: 'Filter on webhooks that have a tag named {name} with a value
          in the given comma-separated list of values.

          The {name} and the value MUST be URL encoded where special characters are
          present.

          Where the tag''s value is a string, at least one of the given values will
          match.

          Where the tag''s value is an array, at least one value in the array will
          match at least one of the given values.

          Partial string matches of the values are not valid.

          '
        schema:
          $ref: vendor/bbc-tams/api/schemas/url-tag-list.json
      - name: tag_exists.{name}
        in: query
        description: 'Filter on webhooks that have a tag named {name} regardless of
          value.

          The {name} MUST be URL encoded where special characters are present.

          If set to true then the presence of the tag is filtered for.

          If set to false then its absence is.

          If left out then no filtering on tag presence is performed.

          '
        schema:
          type: boolean
      - $ref: '#/components/parameters/trait_resource_paged_key'
      - $ref: '#/components/parameters/trait_paged_limit'
      responses:
        '200':
          description: ''
          headers:
            Link:
              description: Provides references to cursors for paging. Only the 'rel'
                attribute with value 'next' and a link to the next page is currently
                supported. If 'next' is not present then it is the last page.
              schema:
                type: string
            X-Paging-Limit:
              description: Identifies the current limit being used for paging. This
                may not match the requested value if the requested value was too high
                for the implementation
              schema:
                type: integer
            X-Paging-NextKey:
              description: Opaque string that can be supplied to the `page` query
                parameter to get the next page of results.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: string
        '404':
          description: Webhooks are not supported by this service implementation
    get:
      summary: List Webhook URLs
      description: 'Get the list of registered webhook URLs.

        Service implementations SHOULD take steps to avoid displaying URLs to users
        other than those who have suitable permissions (e.g. the owning user).

        Availability of this endpoint is indicated by the name "webhooks" appearing
        in the `event_stream_mechanisms` list on the [/service](#/operations/GET_service)
        endpoint.

        '
      operationId: GET_webhooks
      tags:
      - Webhooks
      parameters:
      - name: tag.{name}
        in: query
        description: 'Filter on webhooks that have a tag named {name} with a value
          in the given comma-separated list of values.

          The {name} and the value MUST be URL encoded where special characters are
          present.

          Where the tag''s value is a string, at least one of the given values will
          match.

          Where the tag''s value is an array, at least one value in the array will
          match at least one of the given values.

          Partial string matches of the values are not valid.

          '
        schema:
          $ref: vendor/bbc-tams/api/schemas/url-tag-list.json
      - name: tag_exists.{name}
        in: query
        description: 'Filter on webhooks that have a tag named {name} regardless of
          value.

          The {name} MUST be URL encoded where special characters are present.

          If set to true then the presence of the tag is filtered for.

          If set to false then its absence is.

          If left out then no filtering on tag presence is performed.

          '
        schema:
          type: boolean
      - $ref: '#/components/parameters/trait_resource_paged_key'
      - $ref: '#/components/parameters/trait_paged_limit'
      responses:
        '200':
          description: Return the list of known webhook URLs.
          headers:
            Link:
              description: Provides references to cursors for paging. Only the 'rel'
                attribute with value 'next' and a link to the next page is currently
                supported. If 'next' is not present then it is the last page.
              schema:
                type: string
            X-Paging-Limit:
              description: Identifies the current limit being used for paging. This
                may not match the requested value if the requested value was too high
                for the implementation
              schema:
                type: integer
            X-Paging-NextKey:
              description: Opaque string that can be supplied to the `page` query
                parameter to get the next page of results.
              schema:
                type: string
          content:
            application/json:
              example:
                $ref: vendor/bbc-tams/api/examples/webhook-get-200-list.json
              schema:
                type: array
                items:
                  $ref: vendor/bbc-tams/api/schemas/webhook-get.json
        '404':
          description: Webhooks are not supported by this API implementation
    post:
      summary: Register webhook details
      description: 'Register to receive event notifications as webhooks on a specified
        URL. Webhook messages will conform to the

        format in the `webhooks` section of the API docs, depending on the event type
        (as defined in the same section).

        Availability of this endpoint is indicated by the name "webhooks" appearing
        in the `event_stream_mechanisms`

        list on the service endpoint.


        HTTP requests from the service SHOULD include a `api_key_name` header with
        the ''api_key_value'' value. Clients SHOULD verify this against the value
        they provided when registering the webhook.


        API implementations MAY partially support event filtering and transformations.

        API implementations SHALL return a 400 response code if the filtering or transformation
        specified in the request is not supported.


        API implementations SHOULD consider the security implementations of providing
        webhooks, and include appropriate mitigations against Server Side Request
        Forgery (SSRF) attacks and similar. API implementations SHOULD take appropriate
        steps to authorize the modification of existing webhooks. This may take the
        form of RBAC, or ABAC.

        '
      operationId: POST_webhooks
      tags:
      - Webhooks
      requestBody:
        content:
          application/json:
            example:
              $ref: vendor/bbc-tams/api/examples/webhook-post.json
            schema:
              $ref: vendor/bbc-tams/api/schemas/webhook-post.json
        required: true
      responses:
        '201':
          description: Success. The webhook has been registered.
          content:
            application/json:
              example:
                $ref: vendor/bbc-tams/api/examples/webhook-get-200.json
              schema:
                $ref: vendor/bbc-tams/api/schemas/webhook-get.json
        '400':
          description: Bad request. Invalid parameters or unsupported event filtering
            or transformation.
        '404':
          description: Webhooks are not supported by this service implementation
  /service/webhooks/{webhookId}:
    parameters:
    - name: webhookId
      in: path
      required: true
      schema:
        $ref: vendor/bbc-tams/api/schemas/uuid.json
      description: The Webhook identifier.
    head:
      summary: Webhook details
      description: Return webhook path headers
      operationId: HEAD_webhooks-webhookId
      tags:
      - Webhooks
      responses:
        '200':
          $ref: '#/components/responses/trait_resource_listing_head_200'
        '404':
          description: The requested Webhook ID in the path is invalid, or Webhooks
            are not supported by this service implementation
    get:
      summary: Webhook details
      description: 'Get the details of a webhook. Service implementations SHOULD take
        steps to avoid displaying URLs to users other than those who have suitable
        permissions (e.g. the owning user).

        Availability of this endpoint is indicated by the name "webhooks" appearing
        in the `event_stream_mechanisms` list on the [`/service`](#/operations/GET_service)
        endpoint.

        '
      operationId: GET_webhooks-webhookId
      tags:
      - Webhooks
      responses:
        '200':
          description: Return the webhook details.
          content:
            application/json:
              example:
                $ref: vendor/bbc-tams/api/examples/webhook-get-200.json
              schema:
                $ref: vendor/bbc-tams/api/schemas/webhook-get.json
        '404':
          description: The requested Webhook ID in the path is invalid, or Webhooks
            are not supported by this service implementation
    put:
      summary: Update webhook details
      description: "Update the configuration of an existing webhook.\n\nWebhook messages\
        \ will conform to the format in the `webhooks` section of the API docs, depending\
        \ on the event type (as defined in the same section).\nAvailability of this\
        \ endpoint is indicated by the name \"webhooks\" appearing in the `event_stream_mechanisms`\
        \ list on the [`/service`](#/operations/GET_service) endpoint.\n\nHTTP events\
        \ sent by the service to a client webhook's endpoint SHOULD include a `api_key_name`\
        \ header with the 'api_key_value' value.\nClients SHOULD verify this against\
        \ the value they provided when registering the webhook.\n\nService implementations\
        \ MAY partially support event filtering and transformations.\nService implementations\
        \ SHALL return a 400 response code if the filtering or transformation specified\
        \ in the request is not supported.\n\nService implementations SHOULD consider\
        \ the security implications of providing webhooks, and include appropriate\
        \ mitigations against Server Side Request Forgery (SSRF) attacks and similar.\n\
        Service implementations SHOULD take appropriate steps to authorize the modification\
        \ of existing webhooks. \nThis may take the form of RBAC, or ABAC.\n"
      operationId: PUT_webhooks
      tags:
      - Webhooks
      requestBody:
        content:
          application/json:
            example:
              $ref: vendor/bbc-tams/api/examples/webhook-put.json
            schema:
              $ref: vendor/bbc-tams/api/schemas/webhook-put.json
        required: true
      responses:
        '201':
          description: Success. The webhook has been updated
          content:
            application/json:
              example:
                $ref: vendor/bbc-tams/api/examples/webhook-get-200.json
              schema:
                $ref: vendor/bbc-tams/api/schemas/webhook-get.json
        '400':
          description: Bad request. Invalid parameters or unsupported event filtering
            or transformation.
        '403':
          description: Forbidden. You do not have permission to modify this resource.
        '404':
          description: The requested Webhook ID in the path is invalid, or Webhooks
            are not supported by this service implementation
    delete:
      summary: Delete Webhook
      description: "Deletes the webhook. \nAvailability of this endpoint is indicated\
        \ by the name \"webhooks\" appearing in the `event_stream_mechanisms` list\
        \ on the service endpoint.\n\nService implementations SHOULD consider the\
        \ security implementations of providing webhooks, and include appropriate\
        \ mitigations against Server Side Request Forgery (SSRF) attacks and similar.\
        \ \nService implementations SHOULD take appropriate steps to authorize the\
        \ deleting of webhooks. \nThis may take the form of RBAC, or ABAC.\n"
      operationId: DELETE_webhooks-webhookId
      tags:
      - Webhooks
      responses:
        '204':
          description: No content. The webhook has been deleted.
        '403':
          description: Forbidden. You do not have permission to modify this webhook.
        '404':
          description: The requested Webhook ID in the path is invalid, or Webhooks
            are not supported by this service implementation
...
  /flows/{flowId}/storage:
    parameters:
    - name: flowId
      in: path
      required: true
      schema:
        $ref: vendor/bbc-tams/api/schemas/uuid.json
      description: The Flow identifier.
    post:
      summary: Allocate Initial Flow Storage
      description: 'Allocate initial storage locations for writing Media Objects.


        The Storage Backend type, which is indicated in the [/service](#/operations/GET_service)
        resource, determines the information provided in the response.

        The examples and description below are for the "http_object_store" Storage
        Backend type.

        This Storage Backend type provides HTTP URLs for uploading and downloading
        Media Objects in buckets.


        The response will include a PUT URL that a client uses to upload the Media
        Object.

        The client is expected to register the Flow Segment using the [/flows/{flowId}/segments](#/operations/POST_flows-flowId-segments)
        endpoint once the upload is complete.


        Clients MAY request Objects in batches to reduce the number of HTTP requests
        made to the Service.

        Clients are not expected to use all of the Objects they requested.

        Objects will likely go unused in cases such as shutdown of ingesting clients,
        the end of ingested live streams, and unexpected network congestion.

        Clients SHOULD, however, adapt the number of Objects they request such that
        they may reasonably expect to use them before the timeout advertised in [`min_object_timeout`
        at the `/service`](#/operations/GET_service) endpoint, which is subject to
        a specified minimum (see service endpoint schema).


        Service implementations need to handle situations where Objects are not used,
        and where content was uploaded but no Flow Segment was registered successfully.

        In these circumstances, Services should garbage collect Objects after the
        timeout advertised in [`min_object_timeout` at the `/service`](#/operations/GET_service)
        endpoint.


        When making requests to the provided `put_url`, clients should include credentials
        if the provided URL is on the same origin as the API itself, akin to the `same-origin`
        mode in the [WhatWG Fetch Standard](https://fetch.spec.whatwg.org/#concept-request-credentials-mode).

        '
      operationId: POST_flows-flowId-storage
      tags:
      - MediaStorage
      requestBody:
        content:
          application/json:
            examples:
              limit:
                summary: limit
                externalValue: vendor/bbc-tams/api/examples/flow-storage-post.json
              object_ids:
                summary: object_ids
                externalValue: vendor/bbc-tams/api/examples/flow-storage-post-object_ids.json
            schema:
              $ref: vendor/bbc-tams/api/schemas/flow-storage-post.json
      responses:
        '201':
          description: Storage locations for writing Media Objects.
          content:
            application/json:
              schema:
                $ref: vendor/bbc-tams/api/schemas/flow-storage.json
              examples:
                mpegts:
                  summary: MPEG-TS container Flow
                  description: Flow using MPEG-TS as the container format, writing
                    to a cloud object store
                  externalValue: vendor/bbc-tams/api/examples/flow-storage-post-201.json
                wav_proxy:
                  summary: WAV Flow
                  description: Flow using WAV as the container format, writing to
                    an HTTP interface running on the same origin as the TAMS API
                  externalValue: vendor/bbc-tams/api/examples/flow-storage-post-201-wav.json
        '400':
          description: Bad request. Invalid Flow storage request JSON or the Flow
            'container' is not set. If object_ids supplied, some or all already exist.
        '403':
          description: Forbidden. You do not have permission to modify this Flow.
            It may be marked read-only.
        '404':
          description: The requested Flow does not exist.
...
components:
  responses:
    trait_resource_listing_head_200:
      description: ''
      content:
        application/json:
          schema:
            type: string
...
  parameters:
    trait_paged_limit:
      name: limit
      in: query
      description: Restrict the response to the specified number of results. Service
        implementations may specify their own default and maximum for the limit
      schema:
        type: integer
    trait_resource_paged_key:
      name: page
      in: query
      description: Opaque string used by backend to access a specific page of results.
        Clients should read the next URL from the `Link` header returned with responses,
        or use value of the returned X-Paging-NextKey header. If not supplied, the
        first page is accessed. Service implementations should ensure a consistent
        sort order is applied to pages of results.
      schema:
        type: string
  securitySchemes:
    basic_auth:
      type: http
      scheme: basic
      description: 'HTTP basic authentication using a username and password.


        This option should only be used in constrained cases where Bearer tokens are
        impractical, since the password is intrinsically long-lived, and managing
        user credentials across systems is difficult compared to using centralised
        auth (e.g. via OAuth2) with Bearer tokens

        '
    bearer_token_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'HTTP bearer token authentication.


        The method through which a client should acquire a bearer token is assumed
        to be an OAuth2 grant (either the Authorization Code or Client Credentials
        grant), however the OpenAPI `OAuth Flow` object cannot be used because it
        requires that the relevant URLs are known in advance.


        Clients should provide options to allow one (or both) grant types to be configured,
        as appropriate.


        Where tokens are not JWTs, or are acquired through some other mechanism, the
        API key method should be used instead.

        '
...